Move layers to other file and update functions
